feat: yabloc test
